requirements.txt
requests
firebase-admin
pytz
google-genai